Move tensor extension tests to AIR.
